kanda_tracker/views.py
from rest_framework import viewsets

from .models import Activity, User
from .serializers import ActivitySerializer, UserSerializer


class UserViewSet(viewsets.ModelViewSet):
    queryset = User.objects.all()
    serializer_class = UserSerializer
    permission_classes = []
    filterset_fields = ['id', 'name']


class ActivityViewSet(viewsets.ModelViewSet):
    queryset = Activity.objects.all()
    serializer_class = ActivitySerializer
    permission_classes = []
    filterset_fields = ['id', 'title', 'description']
